requirements.txt
flask
python-decouple
Flask-PyMongo
pymongo[srv]